real-vs-ai-detector/requirements.txt
torch
torchvision
timm
pandas
numpy
scikit-learn
matplotlib
Pillow
streamlit
kagglehub
tqdm
pytest
